long u integer

+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/FotografijaRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.Fotografija;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FotografijaRepository extends JpaRepository<Fotografija, Long> {
+public interface FotografijaRepository extends JpaRepository<Fotografija, Integer> {
 }
