+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/KomentarRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.Komentar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KomentarRepository extends JpaRepository<Komentar, Long> {
+public interface KomentarRepository extends JpaRepository<Komentar, Integer> {
 }
