+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/KorisnikRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.Korisnik;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KorisnikRepository extends JpaRepository<Korisnik, Long> {
+public interface KorisnikRepository extends JpaRepository<Korisnik, Integer> {
 }
