+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/KulturnaPonudaRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.KulturnaPonuda;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KulturnaPonudaRepository extends JpaRepository<KulturnaPonuda, Long> {
+public interface KulturnaPonudaRepository extends JpaRepository<KulturnaPonuda, Integer> {
 }
