+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/OcenaRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.Ocena;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OcenaRepository extends JpaRepository<Ocena, Long> {
+public interface OcenaRepository extends JpaRepository<Ocena, Integer> {
 }
