+++ culturalcontent/src/main/java/com/tim5/culturalcontent/repository/RegistrovaniKorisnikRepository.java
@@ -3,5 +3,5 @@ package com.tim5.culturalcontent.repository;
 import com.tim5.culturalcontent.model.RegistrovaniKorisnik;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegistrovaniKorisnikRepository extends JpaRepository<RegistrovaniKorisnik, Long> {
+public interface RegistrovaniKorisnikRepository extends JpaRepository<RegistrovaniKorisnik, Integer> {
 }
